# daemon/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler_playlistshandler.startRequest:
    properties:
      monitor:
        properties:
          id:
            type: string
          mode:
            $ref: '#/definitions/waypaper-engine_daemon_internal_monitor.MonitorMode'
        type: object
    type: object
...
  waypaper-engine_daemon_internal_handler_httpjson.APIError:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
      error_code:
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
  waypaper-engine_daemon_internal_monitor.MonitorMode:
    enum:
    - individual
    - clone
    - extend
    type: string
    x-enum-varnames:
    - ModeIndividual
    - ModeClone
    - ModeExtend
  waypaper-engine_daemon_internal_store.ActivePlaylistInstance:
    properties:
      current_image_id:
        type: integer
      current_index:
        type: integer
      mode:
        type: string
      monitors:
        items:
          type: string
        type: array
      next_change_at:
        type: string
      next_image_id:
        type: integer
      paused:
        type: boolean
      playlist_id:
        type: integer
      playlist_name:
        type: string
      previous_image_id:
        type: integer
      started_at:
        type: string
      total_images:
        type: integer
    type: object
...
  waypaper-engine_daemon_internal_store.Playlist:
    properties:
      configuration:
        allOf:
        - $ref: '#/definitions/waypaper-engine_daemon_internal_store.PlaylistConfiguration'
        description: Configuration holds the playback behavior settings.
      created_at:
        description: CreatedAt is when the playlist was first created.
        type: string
      id:
        description: ID is the sequential integer identifier, generated daemon-side.
        type: integer
      images:
        description: Images is the ordered list of images in this playlist.
        items:
          $ref: '#/definitions/waypaper-engine_daemon_internal_store.PlaylistImage'
        type: array
      name:
        description: Name is the user-facing display name (e.g. "Evening rotation").
        type: string
      playback:
        allOf:
        - $ref: '#/definitions/waypaper-engine_daemon_internal_store.PlaylistPlayback'
        description: |-
          Playback holds persisted runtime position for daemon restarts. Omitted when
          never started or cleared.
      updated_at:
        description: UpdatedAt is when the playlist was last modified.
        type: string
    type: object
  waypaper-engine_daemon_internal_store.PlaylistConfiguration:
    properties:
      always_start_on_first_image:
        description: AlwaysStartOnFirstImage forces the playlist to start from index
          0 every time.
        type: boolean
      interval:
        description: Interval is the rotation period in seconds (only used for "timer"
          type).
        type: integer
      order:
        description: 'Order is the playback order: "ordered" or "random" (only used
          for "timer" type).'
        type: string
      type:
        description: 'Type is the rotation strategy: "timer", "manual", "time_of_day",
          "day_of_week".'
        type: string
    type: object
  waypaper-engine_daemon_internal_store.PlaylistImage:
    properties:
      image_id:
        description: ImageID references the Image in the gallery.
        type: integer
      media_type:
        description: MediaType is denormalized for quick playlist filtering.
        type: string
      time:
        description: |-
          Time is minutes since midnight (0–1439), used by "time_of_day" playlists.
          Nil for other playlist types.
        type: integer
    type: object
  waypaper-engine_daemon_internal_store.PlaylistPlayback:
    properties:
      current_index:
        type: integer
      mode:
        type: string
      monitors:
        items:
          type: string
        type: array
      paused:
        type: boolean
      timer_cursor:
        type: integer
      timer_indices:
        description: |-
          TimerIndices is the shuffled row order for timer+random; length must match
          images when present.
        items:
          type: integer
        type: array
      was_running:
        type: boolean
    type: object
...
info:
  contact: {}
  description: HTTP API served over a Unix socket for the waypaper-engine wallpaper
    manager.
  title: waypaper-engine daemon
  version: 3.0.0
paths:
...
  /playlists:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/waypaper-engine_daemon_internal_store.Playlist'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: List playlists
      tags:
      - playlists
    post:
      parameters:
      - description: Playlist data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/waypaper-engine_daemon_internal_store.Playlist'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_store.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Create a playlist
      tags:
      - playlists
  /playlists/{id}:
    delete:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Delete a playlist
      tags:
      - playlists
    get:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_store.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get a playlist
      tags:
      - playlists
    patch:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_store.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Update a playlist
      tags:
      - playlists
  /playlists/{id}/next:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Advance to next image in playlist
      tags:
      - playlists
  /playlists/{id}/pause:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Pause a playlist
      tags:
      - playlists
  /playlists/{id}/previous:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Go to previous image in playlist
      tags:
      - playlists
  /playlists/{id}/resume:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Resume a playlist
      tags:
      - playlists
  /playlists/{id}/start:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      - description: Monitor target
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_playlistshandler.startRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Start a playlist
      tags:
      - playlists
  /playlists/{id}/stop:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Stop a playlist
      tags:
      - playlists
...
  /playlists/active/{monitor}:
    get:
      parameters:
      - description: Monitor name
        in: path
        name: monitor
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_store.ActivePlaylistInstance'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get active playlist for a monitor
      tags:
      - playlists
...
swagger: "2.0"